requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
gunicorn==23.0.0
python-dotenv==1.0.1
openpyxl==3.1.5
Werkzeug==3.1.3
fpdf2==2.8.3
Pillow
piexif
openai-whisper
